apps/translation/requirements.txt
google-cloud-storage
google-cloud-translate
uuid